test(cli): split distro contract coverage
